.trash/requirements_backup.lock.txt
--extra-index-url https://download.pytorch.org/whl/cu126



a2wsgi==1.10.8
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
bentoml==1.4.15
binaryornot==0.4.4
cattrs==23.1.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-option-group==0.5.7
clip-interrogator==0.6.0
cloudpickle==3.1.1
coloredlogs==15.0.1
colour-science==0.4.6
comfy-cli==1.4.1
comfyui-embedded-docs==0.2.0
comfyui_frontend_package==1.21.7
comfyui_workflow_templates==0.1.25
contourpy==1.3.2
cookiecutter==2.6.0
cycler==0.12.1
diffusers==0.33.1
dill==0.4.0
duckduckgo_search==8.0.4
easydict==1.13
einops==0.8.1
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.3
frozenlist==1.7.0
fs==2.4.16
fsspec==2025.5.1
ftfy==6.3.1
gdown==5.2.0
gguf==0.17.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kantoku==0.18.3
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lark==1.2.2
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mixpanel==4.10.1
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
open_clip_torch==2.32.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-aiohttp-client==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
packaging==25.0
pathspec==0.12.1
peft==0.15.2
piexif==1.1.3
pillow==11.2.1
pip-requirements-parser==32.0.1
pixeloe==0.1.4
platformdirs==4.3.8
pooch==1.8.2
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
Pygments==2.19.1
PyMatting==1.1.14
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
pyzmq==26.4.0
questionary==2.1.0
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.13
safetensors==0.5.3
schema==0.7.7
scikit-image==0.25.2
scipy==1.15.3
segment-anything==1.0
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
simple-di==0.1.5
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
spandrel==0.4.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
text-unidecode==1.3
tifffile==2025.6.11
timm==1.0.15
tokenizers==0.21.1
tomli_w==1.2.0
tomlkit==0.13.3
torch==2.7.0
torchaudio==2.7.0
torchsde==0.2.6
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.52.4
transparent-background==1.3.4
triton==3.3.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uv==0.7.13
uvicorn==0.34.3
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
wget==3.2
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.30
yarl==1.20.1
zipp==3.23.0
